Add specific school presets for the DE
This is based on https://github.com/openstreetmap/id-tagging-schema/pull/1021

I reused the school and isced level presets and fields from #1021 whenever possible. It will be important to add the correct translations during the translation process
- `school/1-US,DE.json` => "Grundschule"
- `school/2-US,DE.json` => "Sekundarschule ohne Oberstufe"
- `school/2-DE.json` => "Sekundarschule mit Oberstufe"

The ISCED level fields work nicely with the multiCombo by automatically selecting the right preset based on the field combinations.

This is a fresh take on and supersedes https://github.com/openstreetmap/id-tagging-schema/pull/331